Fixed exception when real_idx in metadata is n/a.

+++ backend/main/engines/upload.py
@@ -61,7 +61,7 @@ def get_metadata_from_csv(metadata_file):
     # Clear all rows with all NaN values
     metadata_df.dropna(how="all", inplace=True)
 
-    metadata_df["real_idx"] = metadata_df["real_idx"].astype("int")
+    # metadata_df["real_idx"] = metadata_df["real_idx"].astype("int")
     metadata_df["well1"] = metadata_df["well1"].astype("int")
     metadata_df["well2"] = metadata_df["well2"].astype("int")
     metadata_df["well3"] = metadata_df["well3"].astype("int")
